test: Move all Expect classes to expect.py file
