+++ master/buildbot/test/unit/process/test_buildstep.py
@@ -36,14 +36,14 @@ from buildbot.process.results import RETRY
 from buildbot.process.results import SKIPPED
 from buildbot.process.results import SUCCESS
 from buildbot.process.results import WARNINGS
+from buildbot.test.expect import ExpectGlob
+from buildbot.test.expect import ExpectMkdir
+from buildbot.test.expect import ExpectRmdir
+from buildbot.test.expect import ExpectShell
+from buildbot.test.expect import ExpectStat
 from buildbot.test.fake import fakebuild
 from buildbot.test.fake import fakemaster
 from buildbot.test.fake import worker
-from buildbot.test.fake.remotecommand import ExpectGlob
-from buildbot.test.fake.remotecommand import ExpectMkdir
-from buildbot.test.fake.remotecommand import ExpectRmdir
-from buildbot.test.fake.remotecommand import ExpectShell
-from buildbot.test.fake.remotecommand import ExpectStat
 from buildbot.test.util import config
 from buildbot.test.util import interfaces
 from buildbot.test.util import steps
