+++ master/buildbot/test/unit/steps/test_source_p4.py
@@ -24,8 +24,8 @@ from twisted.trial import unittest
 from buildbot.process.results import RETRY
 from buildbot.process.results import SUCCESS
 from buildbot.steps.source.p4 import P4
-from buildbot.test.fake.remotecommand import ExpectRmdir
-from buildbot.test.fake.remotecommand import ExpectShell
+from buildbot.test.expect import ExpectRmdir
+from buildbot.test.expect import ExpectShell
 from buildbot.test.util import sourcesteps
 from buildbot.test.util.config import ConfigErrorsMixin
 from buildbot.test.util.misc import TestReactorMixin
