+++ master/buildbot/test/unit/steps/test_transfer.py
@@ -32,12 +32,12 @@ from buildbot.process.results import FAILURE
 from buildbot.process.results import SKIPPED
 from buildbot.process.results import SUCCESS
 from buildbot.steps import transfer
-from buildbot.test.fake.remotecommand import ExpectDownloadFile
-from buildbot.test.fake.remotecommand import ExpectGlob
-from buildbot.test.fake.remotecommand import ExpectRemoteRef
-from buildbot.test.fake.remotecommand import ExpectStat
-from buildbot.test.fake.remotecommand import ExpectUploadDirectory
-from buildbot.test.fake.remotecommand import ExpectUploadFile
+from buildbot.test.expect import ExpectDownloadFile
+from buildbot.test.expect import ExpectGlob
+from buildbot.test.expect import ExpectRemoteRef
+from buildbot.test.expect import ExpectStat
+from buildbot.test.expect import ExpectUploadDirectory
+from buildbot.test.expect import ExpectUploadFile
 from buildbot.test.util import steps
 from buildbot.test.util.misc import TestReactorMixin
