+++ master/buildbot/test/unit/test_steps_git_diffinfo.py
@@ -17,7 +17,7 @@ from twisted.trial import unittest
 
 from buildbot.process import results
 from buildbot.steps import gitdiffinfo
-from buildbot.test.fake.remotecommand import ExpectShell
+from buildbot.test.expect import ExpectShell
 from buildbot.test.util import steps
 from buildbot.test.util.misc import TestReactorMixin
